pyproject.toml
[project]
name = "devstoryai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.5.23",
    "crewai>=0.120.1",
    "crewai-tools>=0.45.0",
    "fastapi>=0.115.12",
    "javalang>=0.13.0",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-core>=0.3.60",
    "langchain-google-genai>=2.1.4",
    "langchain-groq>=0.3.7",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "uvicorn>=0.34.2",
]
